Move dot() and matmul() to math.py
