+++ syft/__init__.py
@@ -1,2 +1,2 @@
-from .tensor import TensorBase
 from .tensor import *
+from .math import *
