+++ tests/test_tensor.py
@@ -109,54 +109,6 @@ class EqualTests(unittest.TestCase):
         self.assertTrue(t1 != t2)
 
 
-class DotTests(unittest.TestCase):
-    def testDotInt(self):
-        t1 = TensorBase(np.array([1, 2, 3]))
-        t2 = TensorBase(np.array([4, 5, 6]))
-        self.assertTrue(syft.dot(t1, t2), 32)
-
-    def testDotFloat(self):
-        t1 = TensorBase(np.array([1.3, 2.5, 3.7]))
-        t2 = TensorBase(np.array([4.9, 5.8, 6.5]))
-        self.assertTrue(syft.dot(t1, t2), 43.07)
-
-
-class MatmulTests(unittest.TestCase):
-    def testMatmul1DInt(self):
-        t1 = TensorBase(np.array([1, 2, 3]))
-        t2 = TensorBase(np.array([4, 5, 6]))
-        self.assertTrue(syft.matmul(t1, t2), syft.dot(t1, t2))
-
-    def testMatmul1DFloat(self):
-        t1 = TensorBase(np.array([1.3, 2.5, 3.7]))
-        t2 = TensorBase(np.array([4.9, 5.8, 6.5]))
-        self.assertTrue(syft.matmul(t1, t2), syft.dot(t1, t2))
-
-    def testMatmul2DIdentity(self):
-        t1 = TensorBase(np.array([[1, 0],
-                                  [0, 1]]))
-        t2 = TensorBase(np.array([[5.8, 6.5],
-                                  [7.8, 8.9]]))
-        self.assertTrue(syft.matmul(t1, t2), [[5.8, 6.5],
-                                              [7.8, 8.9]])
-
-    def testMatmul2DInt(self):
-        t1 = TensorBase(np.array([[1.3, 2.5],
-                                  [3.4, 4.5]]))
-        t2 = TensorBase(np.array([[5.8, 6.5],
-                                  [7.8, 8.9]]))
-        self.assertTrue(syft.matmul(t1, t2), [[19, 22],
-                                              [43, 50]])
-
-    def testMatmul2DFloat(self):
-        t1 = TensorBase(np.array([[1.3, 2.5],
-                                  [3.4, 4.5]]))
-        t2 = TensorBase(np.array([[5.8, 6.5],
-                                  [7.8, 8.9]]))
-        self.assertTrue(syft.matmul(t1, t2), [[27.04, 8.45],
-                                              [54.82, 22.1]])
-
-
 def main():
     unittest.main()
